test: add mongo test
