+++ module/data/src/test/kotlin/io/github/siyual_park/data/repository/mongo/MongoRepositoryTestHelper.kt
@@ -1,0 +1,82 @@
+package io.github.siyual_park.data.repository.mongo
+
+import io.github.siyual_park.data.dummy.DummyPerson
+import io.github.siyual_park.data.entity.Person
+import io.github.siyual_park.data.event.BeforeCreateEvent
+import io.github.siyual_park.data.event.BeforeUpdateEvent
+import io.github.siyual_park.data.repository.RepositoryTestHelper
+import io.github.siyual_park.data.repository.mongo.migration.CreatePerson
+import io.github.siyual_park.data.repository.r2dbc.findOneOrFail
+import io.github.siyual_park.event.EventEmitter
+import io.github.siyual_park.event.TypeMatchEventFilter
+import io.github.siyual_park.ulid.ULID
+import kotlinx.coroutines.flow.toList
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Test
+import org.springframework.data.mongodb.core.query.where
+
+abstract class MongoRepositoryTestHelper(
+    repositories: (RepositoryTestHelper<MongoRepository<Person, ULID>>) -> List<MongoRepository<Person, ULID>>,
+) : RepositoryTestHelper<MongoRepository<Person, ULID>>(repositories) {
+    private val eventEmitter = EventEmitter()
+
+    init {
+        eventEmitter.on(TypeMatchEventFilter(BeforeCreateEvent::class), CreateTimestamp())
+        eventEmitter.on(TypeMatchEventFilter(BeforeUpdateEvent::class), UpdateTimestamp())
+    }
+
+    init {
+        migrationManager.register(CreatePerson(mongoTemplate))
+    }
+
+    @Test
+    fun findAllCustomQuery() = parameterized { personRepository ->
+        val person = DummyPerson.create()
+            .let { personRepository.create(it) }
+        val foundPersons = personRepository.findAll(where(Person::id).`is`(person.id)).toList()
+
+        assertEquals(foundPersons.size, 1)
+        assertEquals(person.id, foundPersons[0].id)
+
+        assertEquals(person.name, foundPersons[0].name)
+        assertEquals(person.age, foundPersons[0].age)
+    }
+
+    @Test
+    fun findAllByNameIs() = parameterized { personRepository ->
+        val person = DummyPerson.create()
+            .let { personRepository.create(it) }
+        val foundPersons = personRepository.findAll(where(Person::name).`is`(person.name)).toList()
+
+        assertEquals(foundPersons.size, 1)
+        assertEquals(person.id, foundPersons[0].id)
+
+        assertEquals(person.name, foundPersons[0].name)
+        assertEquals(person.age, foundPersons[0].age)
+    }
+
+    @Test
+    fun findAllByNameIn() = parameterized { personRepository ->
+        val person = DummyPerson.create()
+            .let { personRepository.create(it) }
+        val foundPersons = personRepository.findAll(where(Person::name).`in`(person.name)).toList()
+
+        assertEquals(foundPersons.size, 1)
+        assertEquals(person.id, foundPersons[0].id)
+
+        assertEquals(person.name, foundPersons[0].name)
+        assertEquals(person.age, foundPersons[0].age)
+    }
+
+    @Test
+    fun findOneByName() = parameterized { personRepository ->
+        val person = DummyPerson.create()
+            .let { personRepository.create(it) }
+        val foundPerson = personRepository.findOneOrFail(where(Person::name).`is`(person.name))
+
+        assertEquals(person.id, foundPerson.id)
+
+        assertEquals(person.name, foundPerson.name)
+        assertEquals(person.age, foundPerson.age)
+    }
+}
