+++ module/data/src/test/kotlin/io/github/siyual_park/data/repository/r2dbc/CachedR2DBCRepositoryTest.kt
@@ -1,8 +1,7 @@
-package io.github.siyual_park.data.repository.r2dbc.repository.r2dbc
+package io.github.siyual_park.data.repository.r2dbc
 
 import com.google.common.cache.CacheBuilder
-import io.github.siyual_park.data.repository.r2dbc.R2DBCRepositoryBuilder
-import io.github.siyual_park.data.repository.r2dbc.entity.Person
+import io.github.siyual_park.data.entity.Person
 import io.github.siyual_park.ulid.ULID
 import java.time.Duration
