+++ module/data/src/test/kotlin/io/github/siyual_park/data/repository/r2dbc/SimpleR2DBCRepositoryTest.kt
@@ -1,0 +1,7 @@
+package io.github.siyual_park.data.repository.r2dbc
+
+import io.github.siyual_park.data.entity.Person
+
+class SimpleR2DBCRepositoryTest : R2DBCRepositoryTestHelper(
+    repositories = { listOf(SimpleR2DBCRepository(it.entityOperations, Person::class)) }
+)
