+++ module/data/src/test/kotlin/io/github/siyual_park/data/repository/r2dbc/SoftDeletedR2DBCRepositoryTest.kt
@@ -1,7 +1,6 @@
-package io.github.siyual_park.data.repository.r2dbc.repository.r2dbc
+package io.github.siyual_park.data.repository.r2dbc
 
-import io.github.siyual_park.data.repository.r2dbc.R2DBCRepositoryBuilder
-import io.github.siyual_park.data.repository.r2dbc.entity.Person
+import io.github.siyual_park.data.entity.Person
 import io.github.siyual_park.ulid.ULID
 
 class SoftDeletedR2DBCRepositoryTest : R2DBCRepositoryTestHelper(
